chore: Reorganize all tests
