+++ tests/client/components/allProducts.spec.js
@@ -4,7 +4,7 @@ import {expect} from 'chai';
 import React from 'react';
 import enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import {AllProducts} from './allProducts';
+import {AllProducts} from '../../../client/components/allProducts';
 
 const adapter = new Adapter();
 enzyme.configure({adapter});
