+++ tests/client/components/cart.spec.js
@@ -4,7 +4,7 @@ import {expect} from 'chai';
 import React from 'react';
 import enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import {Cart} from './cart';
+import {Cart} from '../../../client/components/cart';
 
 const adapter = new Adapter();
 enzyme.configure({adapter});
