+++ tests/client/components/singleProduct.spec.js
@@ -4,7 +4,7 @@ import {expect} from 'chai';
 import React from 'react';
 import enzyme, {shallow} from 'enzyme';
 import Adapter from 'enzyme-adapter-react-16';
-import {SingleProduct} from './singleProduct';
+import {SingleProduct} from '../../../client/components/singleProduct';
 
 const adapter = new Adapter();
 enzyme.configure({adapter});
