+++ tests/client/store/user.spec.js
@@ -1,12 +1,12 @@
 /* global describe beforeEach afterEach it */
 
 import {expect} from 'chai';
-import {me, logout} from './user';
+import {me, logout} from '../../../client/store';
 import axios from 'axios';
 import MockAdapter from 'axios-mock-adapter';
 import configureMockStore from 'redux-mock-store';
 import thunkMiddleware from 'redux-thunk';
-import history from '../history';
+import history from '../../../client/history';
 
 const middlewares = [thunkMiddleware];
 const mockStore = configureMockStore(middlewares);
