+++ tests/server/api/products.spec.js
@@ -1,7 +1,7 @@
 const {expect} = require('chai');
 const request = require('supertest');
-const db = require('../db');
-const app = require('../index');
+const db = require('../../../server/db');
+const app = require('../../../server');
 const Product = db.model('product');
 
 describe('GET /products/', () => {
