+++ tests/server/db/product.spec.js
@@ -1,12 +1,12 @@
 /* global describe beforeEach it */
 
 const {expect} = require('chai');
-const db = require('../index');
+const db = require('../../../server/db');
 const Product = db.model('product');
 
 describe('Product model', () => {
 	beforeEach(() => {
-		return db.sync({force: true});
+		return Product.sync({force: true});
 	});
 	describe('create product', () => {
 		it('creates a product with a name, description, price, and imageurl', async () => {
